refactor: splitting codes to another file
PDF-0000
