+++ src/app/api/submit-quiz/route.ts
@@ -1,10 +1,6 @@
 import { NextRequest, NextResponse } from 'next/server'
 import fs from 'fs/promises'
-import OpenAI from 'openai'
-
-const client = new OpenAI({
-	apiKey: process.env.OPEN_AI_API_KEY
-})
+import { client } from '@/lib/open-ai'
 
 export async function POST(req: NextRequest) {
 	const data = await req.json()
