+++ src/app/page.tsx
@@ -2,9 +2,8 @@
 
 import dynamic from 'next/dynamic'
 
-const UploadPDF = dynamic(
-	() =>
-		import('@/components/upload-pdf/upload-pdf').then((mod) => mod.UploadPDF),
+const QuizForm = dynamic(
+	() => import('@/components/quiz-form/quiz-form').then((mod) => mod.QuizForm),
 	{ ssr: false }
 )
 
@@ -12,7 +11,7 @@ export default function Home() {
 	return (
 		<main className='flex flex-col gap-8 items-center p-10'>
 			<h1 className='text-4xl font-bold text-center'>PDF Quiz Generator</h1>
-			<UploadPDF />
+			<QuizForm />
 		</main>
 	)
 }
