+++ src/types/index.ts
@@ -1,0 +1,14 @@
+type Questionnaire = {
+	question: string
+	choices: string[]
+}
+
+type FormValues = {
+	answer1: string
+	answer2: string
+	answer3: string
+	answer4: string
+	answer5: string
+}
+
+export { Questionnaire, FormValues }
